apps/api/requirements.txt
# FastAPI and ASGI Server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database and ORM
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Google API clients
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0

# JWT Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and Configuration
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Development and Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Background Tasks (for data ingestion)
celery==5.3.6
redis==5.0.7

# AI/ML Integration (for future LangGraph integration)
langchain==0.2.5
langsmith==0.1.77